changed evaluation from floats to ints using centipawns

+++ header_files/engine.h
@@ -7,14 +7,14 @@
 #include "../header_files/move.h"
 #include "../header_files/inout.h"
 
-float get_best_move_minimax(Board* board, Move* best_move, int max_depth);
-float maxi(Board* board, int depth, float alpha, float beta);
-float mini(Board* board, int depth, float alpha, float beta);
+int get_best_move_minimax(Board* board, Move* best_move, int max_depth);
+int maxi(Board* board, int depth, int alpha, int beta);
+int mini(Board* board, int depth, int alpha, int beta);
 
-float evaluate(Board* board);
+int evaluate(Board* board);
 
-float calculate_material_score(Board* board);
-float get_piece_value(int piece_type);
+int calculate_material_score(Board* board);
+int get_piece_value(int piece_type);
 
-float calculate_movement_score(Board* board);
+int calculate_movement_score(Board* board);
